+++ BambooTracker/lang/bamboo_tracker_fr.ts
@@ -258,7 +258,7 @@
     </message>
     <message>
         <location filename="../gui/configuration_dialog.ui" line="207"/>
-        <source>Move cursor by horizontal scroll</source>
+        <source>Move cursor with horizontal scroll bar</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -713,13 +713,13 @@
     </message>
     <message>
         <location filename="../gui/configuration_dialog.ui" line="111"/>
-        <source>Warp cursor</source>
+        <source>Wrap cursor</source>
         <translatorcomment>J&apos;ai rien trouvé de mieux.</translatorcomment>
         <translation type="unfinished">Déplacer le curseur de manière cyclique</translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.ui" line="119"/>
-        <source>Warp across orders</source>
+        <source>Wrap across orders</source>
         <translatorcomment>J&apos;ai rien trouvé de mieux. Incertitude à propos de la traduction adéquate pour &quot;order&quot;</translatorcomment>
         <translation type="unfinished">Se déplacer de manière cyclique le long des ordres</translation>
     </message>
@@ -751,7 +751,7 @@
     </message>
     <message>
         <location filename="../gui/configuration_dialog.ui" line="167"/>
-        <source>Move cursor to right</source>
+        <source>Move cursor right</source>
         <translation>Déplacer le curseur vers la droite</translation>
     </message>
     <message>
@@ -775,7 +775,7 @@
     </message>
     <message>
         <location filename="../gui/configuration_dialog.ui" line="215"/>
-        <source>Overwrite unused&amp;unedited property</source>
+        <source>Overwrite unused and unedited properties</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -917,13 +917,13 @@
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="103"/>
-        <source>Warp the cursor around the edges of the pattern editor.</source>
+        <source>Wrap the cursor around the edges of the pattern editor.</source>
         <translatorcomment>J&apos;ai rien trouvé de mieux.</translatorcomment>
         <translation type="unfinished">Le curseur se déplace de manière cyclique à proximité des bords de l&apos;éditeur de motifs.</translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="105"/>
-        <source>Move to previous or next order when reaching top or bottom in the pattern editor.</source>
+        <source>Move to the previous or next order when reaching the top or bottom in the pattern editor.</source>
         <translatorcomment>Incertitude à propos de la traduction adéquate pour &quot;order&quot;</translatorcomment>
         <translation type="unfinished">Se positionner sur l&apos;ordre précédent ou suivant lorsque le sommet ou le bas de l&apos;éditeur de motifs est atteint.</translation>
     </message>
@@ -939,7 +939,7 @@
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="109"/>
-        <source>Preview previous and next orders in the pattern editor.</source>
+        <source>Preview the previous and next orders in the pattern editor.</source>
         <translatorcomment>Incertitude à propos de la traduction adéquate pour &quot;order&quot;</translatorcomment>
         <translation type="unfinished">Prévisualiser les ordres précédent et suivant dans l&apos;éditeur de motifs.</translation>
     </message>
@@ -960,7 +960,7 @@
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="117"/>
-        <source>Move the cursor to right after entering effects in the pattern editor.</source>
+        <source>Automatically move the cursor right after entering effects in the pattern editor.</source>
         <translation>Déplacer le curseur vers la droite après être entré dans les effets dans l&apos;éditeur de motif.</translation>
     </message>
     <message>
@@ -984,37 +984,37 @@
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="125"/>
-        <source>Fill 00 to effect value column upon entering effect id.</source>
+        <source>Automatically fill 00 to the effect value column upon entering an effect ID.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="127"/>
-        <source>Move the cursor position by cell with horizontal scroll bar in the order list and the pattern editor.</source>
+        <source>Move the cursor with the horizontal scroll bar in the order list and the pattern editor.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="129"/>
-        <source>Overwrite unused and unedited instrument properties on creating new properties. When disabled, override unused properties regardless of editing.</source>
+        <source>Overwrite unused and unedited instrument properties when creating new properties. If disabled, override unused properties regardless of editing.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="132"/>
-        <source>Send only ADPCM samples used by instruments to the ADPCM memory. Recommend to turn off if you change ADPCM samples frequently due to take the high rewriting cost.</source>
+        <source>Only send samples used by instruments to the ADPCM memory. It is recommended to turn this off if you change ADPCM samples frequently due to the high cost of rewriting..</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="135"/>
-        <source>Correspond the instrument number in patterns when the instrument changes its number.</source>
+        <source>Automatically update the instrument number in patterns when an instrument's number is changed.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="137"/>
-        <source>Set maximum volume during jam mode. When unchecked, the volume is changed by the volume spinbox.</source>
+        <source>Set maximum volume during jam mode. When unchecked, the volume is changed by the volume combo box.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="139"/>
-        <source>Mute hidden tracks when visibility of tracks is changed.</source>
+        <source>Automatically mute tracks when they are hidden.</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
@@ -1184,22 +1184,22 @@
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="200"/>
-        <source>Transpose, decrease note</source>
+        <source>Transpose down one semitone</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="201"/>
-        <source>Transpose, increase note</source>
+        <source>Transpose up one semitone</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="202"/>
-        <source>Transpose, decrease octave</source>
+        <source>Transpose down one octave</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="../gui/configuration_dialog.cpp" line="203"/>
-        <source>Transpose, increase octave</source>
+        <source>Transpose up one octave</source>
         <translation type="unfinished"></translation>
     </message>
     <message>